Started working on making the album sortable. All code responsible for viewing the album will now be in SortablePhotoAlbum.js and PhotoViewer.js will be responsible for preparing the photos
