+++ package.json
@@ -13,6 +13,10 @@
     "url": "git+https://github.com/creativetimofficial/ct-material-kit-react.git"
   },
   "dependencies": {
+    "@dnd-kit/core": "^6.3.1",
+    "@dnd-kit/modifiers": "^9.0.0",
+    "@dnd-kit/sortable": "^10.0.0",
+    "@dnd-kit/utilities": "^3.2.2",
     "@emotion/cache": "11.10.7",
     "@emotion/react": "^11.14.0",
     "@emotion/styled": "^11.14.0",
@@ -25,6 +29,7 @@
     "@testing-library/react": "14.0.0",
     "@testing-library/user-event": "14.4.3",
     "@toolpad/core": "^0.15.0",
+    "array-move": "^4.0.0",
     "chroma-js": "2.4.2",
     "flatpickr": "4.6.13",
     "prop-types": "15.8.1",
@@ -37,7 +42,7 @@
     "react-icons": "^5.5.0",
     "react-imgix": "^9.10.0",
     "react-loader-spinner": "^6.1.6",
-    "react-photo-album": "^3.0.2",
+    "react-photo-album": "^3.1.0",
     "react-router-dom": "6.10.0",
     "react-router-hash-link": "^2.4.3",
     "react-scripts": "5.0.1",
