+++ src/components/SortablePhotoAlbum/admin/EditView.js
