+++ src/components/SortablePhotoAlbum/admin/SelectIcon.js
