+++ src/components/SortablePhotoAlbum/admin/SelectIcon.module.css
