+++ src/components/SortablePhotoAlbum/admin/StyledLink.js
